moved concat_video function to video_utils, cleaned some code
